Updating the headers

+++ headers/CompElementTemplate.h
@@ -69,7 +69,7 @@ public:
     virtual int64_t GetDOFIndex(int i) const;
     
     // Return the number of degree of freedom
-    virtual int NDOF() const;
+    virtual int64_t NDOF() const;
     
     // Return the number of shape functions stored in the DOF data structure
     virtual int NShapeFunctions(int doflocindex) const;
